feat: move nft specific functions to outside of VWBLBase
