+++ src/vwbl/erc721/VWBL.ts
@@ -1,10 +1,20 @@
+import axios from "axios";
 import * as fs from "fs";
 
 import { uploadEncryptedFile, uploadMetadata, uploadThumbnail } from "../../storage/aws/upload";
-import { createRandomKey, encryptFile, encryptStream, encryptString } from "../../util/cryptoHelper";
+import {
+  createRandomKey,
+  decryptFile,
+  decryptStream,
+  decryptString,
+  encryptFile,
+  encryptStream,
+  encryptString,
+} from "../../util/cryptoHelper";
 import { getMimeType, toBase64FromBlob } from "../../util/fileHelper";
 import { ConstructorProps, VWBLBase } from "../base";
 import { VWBLNFT } from "../blockchain";
+import { ExtractMetadata, Metadata, PlainMetadata } from "../metadata";
 import {
   EncryptLogic,
   UploadContentType,
@@ -14,7 +24,9 @@ import {
   UploadThumbnail,
 } from "../types";
 
-export class VWBL extends VWBLBase<VWBLNFT> {
+export class VWBL extends VWBLBase {
+  nft: VWBLNFT;
+
   constructor(props: ConstructorProps) {
     super(props);
 
@@ -183,6 +195,23 @@ export class VWBL extends VWBLBase<VWBLNFT> {
   };
 
   /**
+   * Set key to VWBL Network
+   *
+   * @param tokenId - The ID of NFT
+   * @param key - The key generated by {@link VWBL.createKey}
+   * @param hasNonce
+   * @param autoMigration
+   *
+   */
+  setKey = async (tokenId: number, key: string, hasNonce?: boolean, autoMigration?: boolean): Promise<void> => {
+    if (!this.signature) {
+      throw "please sign first";
+    }
+    const { documentId } = await this.nft.getTokenInfo(tokenId);
+    return await this._setKey(documentId, key, hasNonce, autoMigration);
+  };
+
+  /**
    * Mint new NFT
    *
    * @param royaltiesPercentage - This percentage of the sale price will be paid to the NFT creator every time the NFT is sold or re-sold
@@ -222,5 +251,215 @@ export class VWBL extends VWBLBase<VWBLNFT> {
    */
   safeTransfer = async (to: string, tokenId: number): Promise<void> => {
     await this.nft.safeTransfer(to, tokenId);
+  };
+
+  /**
+   * Uplod Metadata
+   *
+   * @remarks
+   * By default, metadata will be uploaded to Amazon S3.
+   * You need to pass `uploadMetadataCallBack` if you upload metadata to a storage other than Amazon S3.
+   *
+   * @param tokenId - The ID of NFT
+   * @param name - The NFT name
+   * @param description - The NFT description
+   * @param thumbnailImageUrl - The URL of the thumbnail image
+   * @param encryptedDataUrls - The URL of the encrypted file data
+   * @param mimeType - The mime type of encrypted file data
+   * @param encryptLogic - Select ether "base64" or "binary". Selection criteria: "base64" -> sutable for small data. "binary" -> sutable for large data.
+   * @param uploadMetadataCallBack - Optional: the function for uploading metadata
+   */
+  uploadMetadata = async (
+    tokenId: number,
+    name: string,
+    description: string,
+    thumbnailImageUrl: string,
+    encryptedDataUrls: string[],
+    mimeType: string,
+    encryptLogic: EncryptLogic,
+    uploadMetadataCallBack?: UploadMetadata
+  ): Promise<void> => {
+    const { uploadMetadataType, awsConfig } = this.opts;
+    const uploadMetadataFunction =
+      uploadMetadataType === UploadMetadataType.S3 ? uploadMetadata : uploadMetadataCallBack;
+    if (!uploadMetadataFunction) {
+      throw new Error("please specify upload metadata type or give callback");
+    }
+    await uploadMetadataFunction(
+      tokenId,
+      name,
+      description,
+      thumbnailImageUrl,
+      encryptedDataUrls,
+      mimeType,
+      encryptLogic,
+      awsConfig
+    );
+  };
+
+  /**
+   * Uplod Metadata to IPFS
+   *
+   * @remarks
+   * Metadata will be uploaded to IPFS.
+   *
+   * @param tokenId - The ID of NFT
+   * @param name - The NFT name
+   * @param description - The NFT description
+   * @param thumbnailImageUrl - The URL of the thumbnail image
+   * @param encryptedDataUrls - The URL of the encrypted file data
+   * @param mimeType - The mime type of encrypted file data
+   * @param encryptLogic - Select ether "base64" or "binary". Selection criteria: "base64" -> sutable for small data. "binary" -> sutable for large data.
+   */
+  uploadMetadataToIPFS = async (
+    name: string,
+    description: string,
+    thumbnailImageUrl: string,
+    encryptedDataUrls: string[],
+    mimeType: string,
+    encryptLogic: EncryptLogic
+  ): Promise<string> => {
+    const metadataUrl = await this.uploadToIpfs?.uploadMetadata(
+      name,
+      description,
+      thumbnailImageUrl,
+      encryptedDataUrls,
+      mimeType,
+      encryptLogic
+    );
+    return metadataUrl as string;
+  };
+
+  /**
+   * Get all NFT metadata owned by a person who call this method
+   *
+   * @returns Array of token metadata
+   */
+  getOwnTokens = async (): Promise<Metadata[]> => {
+    if (!this.signature) {
+      throw "please sign first";
+    }
+    const ownTokenIds = await this.nft.getOwnTokenIds();
+    return (await Promise.all(ownTokenIds.map(this.getMetadata.bind(this)))).filter(
+      (extractMetadata): extractMetadata is Metadata => extractMetadata !== undefined
+    );
+  };
+
+  /**
+   * Get token IDs owned by someone who called this function
+   *
+   * @returns Array of token IDs
+   */
+  getOwnTokenIds = async (): Promise<number[]> => {
+    return await this.nft.getOwnTokenIds();
+  };
+
+  /**
+   * Get NFT metadata from given `tokenId`
+   *
+   * @remarks
+   * Check if a person call this method is a NFT owner, and if so, return a decrypted data.
+   *
+   * @param tokenId - The ID of NFT
+   * @returns Token metadata and an address of NFT owner
+   */
+  getTokenById = async (tokenId: number): Promise<(ExtractMetadata | Metadata) & { owner: string }> => {
+    const isOwnerOrMinter = (await this.nft.isOwnerOf(tokenId)) || (await this.nft.isMinterOf(tokenId));
+    const owner = await this.nft.getOwner(tokenId);
+    const metadata = isOwnerOrMinter ? await this.extractMetadata(tokenId) : await this.getMetadata(tokenId);
+    if (!metadata) {
+      throw new Error("metadata not found");
+    }
+    return { ...metadata, owner };
+  };
+
+  /**
+   * Get token ids by minter address
+   * @param address - minter address
+   * @returns Token ids
+   */
+  getTokenByMinter = async (address: string): Promise<number[]> => {
+    return await this.nft.getTokenByMinter(address);
+  };
+
+  /**
+   * Get NFT metadata from given `tokenId`
+   *
+   * @param tokenId - The ID of NFT
+   * @returns Token metadata
+   */
+  getMetadata = async (tokenId: number): Promise<Metadata | undefined> => {
+    const metadataUrl = await this.nft.getMetadataUrl(tokenId);
+    const metadata = (await axios.get(metadataUrl).catch(() => undefined))?.data;
+    // delete token if metadata is not found
+    if (!metadata) {
+      return undefined;
+    }
+    return {
+      id: tokenId,
+      name: metadata.name,
+      description: metadata.description,
+      image: metadata.image,
+      mimeType: metadata.mime_type,
+      encryptLogic: metadata.encrypt_logic,
+    };
+  };
+
+  /**
+   * Get NFT metadata from given `tokenId`
+   *
+   * @remarks
+   * This method should be called by NFT owner.
+   *
+   * @param tokenId The ID of NFT
+   * @returns Token metadata
+   */
+  extractMetadata = async (tokenId: number): Promise<ExtractMetadata | undefined> => {
+    if (!this.signature) {
+      throw "please sign first";
+    }
+    const metadataUrl = await this.nft.getMetadataUrl(tokenId);
+    const metadata: PlainMetadata = (await axios.get(metadataUrl).catch(() => undefined))?.data;
+    // delete token if metadata is not found
+    if (!metadata) {
+      return undefined;
+    }
+    const { documentId } = await this.nft.getTokenInfo(tokenId);
+    const chainId = await this.opts.web3.eth.getChainId();
+    const decryptKey = await this.api.getKey(documentId, chainId, this.signature);
+    const encryptedDataUrls = metadata.encrypted_data;
+    const isRunningOnBrowser = typeof window !== "undefined";
+    const encryptLogic = metadata.encrypt_logic ?? "base64";
+    const ownDataArray = await Promise.all(
+      encryptedDataUrls.map(async (encryptedDataUrl) => {
+        const encryptedData = (
+          await axios.get(encryptedDataUrl, {
+            responseType: encryptLogic === "base64" ? "text" : isRunningOnBrowser ? "arraybuffer" : "stream",
+          })
+        ).data;
+        return encryptLogic === "base64"
+          ? decryptString(encryptedData, decryptKey)
+          : isRunningOnBrowser
+          ? await decryptFile(encryptedData, decryptKey)
+          : decryptStream(encryptedData, decryptKey);
+      })
+    );
+    const ownFiles = ownDataArray.filter((ownData): ownData is ArrayBuffer => typeof ownData !== "string");
+    const ownDataBase64 = ownDataArray.filter((ownData): ownData is string => typeof ownData === "string");
+    const fileName = encryptedDataUrls[0]
+      .split("/")
+      .slice(-1)[0]
+      .replace(/\.vwbl/, "");
+    return {
+      id: tokenId,
+      name: metadata.name,
+      description: metadata.description,
+      image: metadata.image,
+      mimeType: metadata.mime_type,
+      encryptLogic: metadata.encrypt_logic,
+      ownDataBase64,
+      ownFiles,
+      fileName,
+    };
   };
 }
